requirements.txt
absl-py==1.0.0
aiofiles==23.2.1
aiosignal==1.2.0
alembic==1.10.3
anyio==4.0.0
asttokens==2.2.1
astunparse==1.6.3
attrs==20.3.0
autograd==1.4
autopep8==2.0.0
backcall==0.2.0
backpack-for-pytorch==1.6.0
black==22.10.0
boto3==1.24.16
botocore==1.27.16
brotlipy==0.7.0
cachetools==5.1.0
certifi==2022.5.18.1
cffi
charset-normalizer==2.0.12
click==8.0.4
cmaes==0.9.1
cmake==3.26.3
colorlog==6.7.0
comm==0.1.2
contextlib2==21.6.0
cramjam==2.5.0
cryptography
cycler==0.11.0
debugpy==1.6.5
decorator==5.1.1
dill==0.3.5.1
distlib==0.3.4
einops==0.6.1
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.1.3
executing==1.2.0
fastparquet==0.8.1
filelock==3.7.1
flatbuffers==23.5.26
fonttools==4.33.3
frozenlist==1.3.0
fsspec==2022.5.0
future==0.18.2
gast==0.4.0
gmpy2
google-auth==2.19.1
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
graph-onedrive==0.4.0
greenlet==2.0.2
grpcio==1.54.2
h11==0.14.0
h5py==3.7.0
httpcore==0.18.0
httpx==0.25.0
idna==3.3
imageio==2.27.0
importlib-metadata==4.11.4
iniconfig==1.1.1
ipykernel==6.20.2
ipython==8.8.0
jax==0.4.11
jedi==0.18.2
Jinja2
jmespath==1.0.1
joblib==1.1.0
jsonschema==4.6.0
jupyter_client==7.4.9
jupyter_core==5.1.3
keras==2.12.0
kiwisolver==1.4.2
lazy_loader==0.2
libclang==16.0.0
lightning-utilities==0.9.0
lit==16.0.6
littleutils==0.2.2
Mako==1.2.4
Markdown==3.3.7
MarkupSafe==2.1.2
matplotlib==3.5.2
matplotlib-inline==0.1.6
ml-dtypes==0.1.0
mpmath==1.3.0
msgpack==1.0.4
multiprocess==0.70.13
mypy-extensions==0.4.3
natsort==8.4.0
nest-asyncio==1.5.6
networkx==3.1
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.0
ogb==1.3.6
opencv-python==4.7.0.72
openpyxl==3.1.2
opt-einsum==3.3.0
optuna==3.1.1
outdated==0.2.2
packaging==21.3
pandas==1.3.5
parso==0.8.3
pathos==0.2.9
pathspec==0.10.2
patsy==0.5.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.1
platformdirs==2.5.2
plotly==5.17.0
pluggy==1.0.0
pox==0.3.1
ppft==1.7.6.5
prompt-toolkit==3.0.36
protobuf==3.20.3
protobuf3-to-dict==0.1.5
psutil==5.9.4
ptflops==0.6.9
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pyaml==21.10.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pycparser
Pygments==2.14.0
pyOpenSSL
pyparsing==3.0.9
pyrsistent==0.18.1
PySocks
pytest==7.1.2
python-dateutil==2.8.2
pytz==2022.1
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.0.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
s3fs==0.4.2
s3transfer==0.6.0
sagemaker==2.117.0
schema==0.7.5
SciencePlots==2.0.1
scikit-image==0.20.0
scikit-learn==1.1.1
scikit-optimize==0.9.0
scipy==1.9.1
seaborn==0.12.2
Shapely==1.8.5.post1
six
sklearn==0.0
smdebug-rulesconfig==1.0.1
sniffio==1.3.0
SQLAlchemy==2.0.9
stack-data==0.6.2
statsmodels==0.13.2
sympy==1.12
tabulate==0.8.10
tenacity==8.2.3
tensorboard==2.12.3
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.3.0
threadpoolctl==3.1.0
tifffile==2023.4.12
tomli==2.0.1
torch==2.0.1
torch-tb-profiler==0.4.0
torchaudio==2.0.2
torchmetrics==1.2.0
torchvision==0.15.2
tornado==6.2
tqdm==4.64.0
traitlets==5.8.1
triton==2.0.0
typing_extensions==4.2.0
ujson==5.3.0
uncertainty_toolbox==0.1.0
unfoldNd==0.2.0
urllib3==1.26.9
virtualenv==20.14.1
wcwidth==0.2.6
Werkzeug==2.1.2
wilds==2.0.0
wrapt==1.14.1
xlrd==2.0.1
zipp==3.8.0